permission with project

+++ src/dthaus/projects/views.py
@@ -12,6 +12,7 @@ from .forms import ProjectCreateForm, PhaseCreateForm, ListPhaseForm, \
 
 brand = 'DONGJIN VIETNAM J.S.C'
 
+
 def all_permission(request):
     permission = []
 
@@ -23,11 +24,12 @@ def all_permission(request):
                 data['object_type'] = group.object_type
                 data['object_id'] = group.object_id
                 data['permission'] = per
-                
+
                 if data not in permission:
                     permission.append(data)
-    
+
     return permission
+
 
 def has_permission(request, object_type, object_id, permission):
     if request.user.user_group.all():
@@ -36,8 +38,9 @@ def has_permission(request, object_type, object_id, permission):
             for per in per_group:
                 if group.object_type == object_type and str(group.object_id) == str(object_id) and per.codename == permission:
                     return True
-    
+
     return False
+
 
 @login_required(login_url='login')
 def projects(request):
@@ -64,30 +67,25 @@ def projects(request):
                     project = Project.objects.get(id=permission['object_id'])
                     if project not in projects:
                         projects.append(project)
-            
-                        for phase in Phase.objects.all():
-                            # Get all phase of project
-                            if phase.project == project:
-                                phases.append(phase)
-                                
-                                for task in Task.objects.all():
-                                    if task.phase == phase:
-                                        tasks.append(task)
+
+                        for phase in Phase.objects.filter(project=project):
+                            phases.append(phase)
+                            for task in Task.objects.filter(phase=phase):
+                                tasks.append(task)
                 # Permission view phase
                 if permission['object_type'] == 'phase':
                     phase = Phase.objects.get(id=permission['object_id'])
                     if phase not in phases:
                         phases.append(phase)
 
-                        for task in Task.objects.all():
-                            if task.phase == phase:
-                                tasks.append(task)
+                        for task in Task.objects.filter(phase=phase):
+                            tasks.append(task)
                 # Permission view task
                 if permission['object_type'] == 'task':
                     task = Task.objects.get(id=permission['object_id'])
                     if task not in tasks:
                         tasks.append(task)
-            
+
             # Permission
             if str(permission['permission']) == 'add':
                 project_per_add = True
@@ -96,17 +94,12 @@ def projects(request):
             if str(permission['permission']) == 'delete':
                 project_per_delete = True
 
-    print('project_per_add', project_per_add)
-    print('project_per_change', project_per_change)
-    print('project_per_delete', project_per_delete)
-
     context = {
         'brand': brand,
         'title': title,
         'projects': projects,
         'phases': phases,
         'tasks': tasks,
-        'all_permission': all_permission(request),
         'project_per_add': project_per_add,
         'project_per_change': project_per_change,
         'project_per_delete': project_per_delete,
@@ -119,21 +112,33 @@ def projects(request):
 def project_create(request):
     title = 'Create Project'
     stage = 1
-    create_form = ProjectCreateForm()
-    list_phase = CustomPhase.objects.all()
+    project_per_add = False
 
-    if request.method == 'POST':
-        create_form = ProjectCreateForm(request.POST)
-        if create_form.is_valid():
-            new_project = create_form.save()
-            return redirect('phase_settings', pk_project=new_project.id)
+    for permission in all_permission(request):
+        # Permission add project
+        if str(permission['permission']) == 'add':
+            project_per_add = True
+            break
+
+    create_form = ''
+    if project_per_add or request.user.is_superuser:
+        create_form = ProjectCreateForm()
+
+        if request.method == 'POST':
+            create_form = ProjectCreateForm(request.POST)
+            if create_form.is_valid():
+                new_project = create_form.save()
+                return redirect('phase_settings', pk_project=new_project.id)
+    else:
+        messages.add_message(request, messages.ERROR,
+                             "You don't have permissions for this action.")
+        return redirect('projects')
 
     context = {
         'brand': brand,
         'title': title,
         'stage': stage,
         'create_form': create_form,
-        'list_phase': list_phase,
     }
 
     return render(request, 'projects/project-create.html', context=context)
@@ -142,24 +147,59 @@ def project_create(request):
 @login_required(login_url='login')
 def project_update(request, pk_project):
     title = 'Update Project'
-    project = Project.objects.get(id=pk_project)
-    phases = Phase.objects.filter(project=pk_project)
-    project_update_form = ProjectCreateForm(instance=project)
-    PhaseFormSet = inlineformset_factory(Project, Phase,
-                                         PhaseCreateForm,
-                                         fields=(
-                                             'phase_name', 'start_date',
-                                             'due_date', 'phase_status',
-                                             'phase_finished', 'project',
-                                         ), max_num=len(phases))
-    phase_form = PhaseFormSet(instance=project)
+    projects = []
+    phases = []
+    tasks = []
+    project_per_change = False
 
-    if request.method == 'POST':
-        # print('POST', request.POST)
-        project_update_form = ProjectCreateForm(request.POST, instance=project)
-        if project_update_form.is_valid():
-            project_update_form.save()
-            return redirect('projects')
+    for permission in all_permission(request):
+        # Permission view project
+        if str(permission['permission']) == 'view':
+            if permission['object_type'] == 'project':
+                project = Project.objects.get(id=permission['object_id'])
+                if project not in projects:
+                    projects.append(project)
+
+                    for phase in Phase.objects.filter(project=project):
+                        phases.append(phase)
+
+                        for task in Task.objects.filter(phase=phase):
+                            tasks.append(task)
+
+        # Permission change project
+        if str(permission['permission']) == 'change':
+            project_per_change = True
+
+    project = Project.objects.get(id=pk_project)
+    project_update_form = ''
+    phase_form = ''
+    # Check permission change project
+    if project in projects and project_per_change or request.user.is_superuser:
+        phases = Phase.objects.filter(project=pk_project)
+        project_update_form = ProjectCreateForm(instance=project)
+
+        PhaseFormSet = inlineformset_factory(Project, Phase,
+                                             PhaseCreateForm,
+                                             fields=(
+                                                 'name', 'start_date',
+                                                 'due_date', 'phase_status',
+                                                 'phase_finished', 'project',
+                                             ), max_num=len(phases))
+        phase_form = PhaseFormSet(instance=project)
+
+        if request.method == 'POST':
+            # print('POST', request.POST)
+            project_update_form = ProjectCreateForm(
+                request.POST, instance=project)
+            if project_update_form.is_valid():
+                project_update_form.save()
+                messages.add_message(
+                    request, messages.SUCCESS, "Project was updated.")
+                return redirect('projects')
+    else:
+        messages.add_message(request, messages.ERROR,
+                             "You don't have permissions for this action.")
+        return redirect("projects")
 
     context = {
         'brand': brand,
@@ -174,11 +214,30 @@ def project_update(request, pk_project):
 @login_required(login_url='login')
 def project_delete(request, pk_project):
     title = 'Delete Project'
-    project = Project.objects.get(id=pk_project)
+    projects = []
+    project_per_delete = False
 
-    if request.method == 'POST':
-        project.delete()
-        return redirect('projects')
+    for permission in all_permission(request):
+        # Permission view project
+        if str(permission['permission']) == 'view':
+            if permission['object_type'] == 'project':
+                project = Project.objects.get(id=permission['object_id'])
+                if project not in projects:
+                    projects.append(project)
+
+        # Permission change project
+        if str(permission['permission']) == 'delete':
+            project_per_delete = True
+
+    project = Project.objects.get(id=pk_project)
+    if project in projects and project_per_delete or request.user.is_superuser:
+        if request.method == 'POST':
+            project.delete()
+            return redirect('projects')
+    else:
+        messages.add_message(request, messages.ERROR,
+                             "You don't have permissions for this action.")
+        return redirect("projects")
 
     context = {
         'brand': brand,
@@ -221,12 +280,12 @@ def phase_settings(request, pk_project):
             for phase_item in list_phase:
                 if phase_item.selected:
                     phase_exists = Phase.objects.filter(
-                        phase_name=phase_item.phase_name, project=project)
+                        name=phase_item.phase_name, project=project)
                     # Check phase name exist in table Phase of Project
 
                     if not phase_exists:
                         new_phase = Phase(
-                            phase_name=phase_item.phase_name, project=project)
+                            name=phase_item.phase_name, project=project)
                         try:
                             new_phase.save()
                             messages.add_message(
@@ -433,11 +492,11 @@ def task_edits(request, pk_project, pk_phase, pk_task):
     # Get attachment of Task
     task_files = TaskFiles.objects.filter(task=pk_task)
     last_file = task_files[len(task_files)-1] if len(task_files) > 0 else 0
-    
+
     file_formset = inlineformset_factory(Task, TaskFiles,
-                                        FileApproveForm,
-                                        fields='__all__', 
-                                        max_num=len(task_files))
+                                         FileApproveForm,
+                                         fields='__all__',
+                                         max_num=len(task_files))
     approve_form = file_formset(instance=task)
     if request.method == 'POST':
         if 'btn-approve-file' in request.POST:
@@ -452,7 +511,8 @@ def task_edits(request, pk_project, pk_phase, pk_task):
             else:
                 print(form.errors)
 
-            messages.add_message(request, messages.SUCCESS, 'Approve file successfully.')
+            messages.add_message(request, messages.SUCCESS,
+                                 'Approve file successfully.')
             return redirect('task_edits', project.id, phase.id, task.id)
 
     context = {
